Remove commented lines in CMakeLists.txt
There is an issue with running non-Debug builds in the webpack step so this change allows the build to succeed.  The issue is between the src/web/wepback.dev.in and webpack.prod.in files.  We have opened an issue in the RKF Jira (RAT-293) but that should perhaps transition to the TIP Jira