OA-393. Variable declaration shadows another GdalNameMapperDirect constructor opened/closed GDAL file in try block using ‘GDALDataset *gdalDataSet’ variable. Catch clause of this block closed GDAL file if exception happens. However it used different instance of this variable, declared at higher-level scope.
Hence exception would lead to dangling handle to GDAL file, i.e. resource leak. Not a big deal now, as such exception will terminate the program, still it must be fixed.

This problem was not revealed earlier because of an old GCC version being used - it does not print warning on this problem even on -Wall -Wextra -Wpedantic level.